+++ tests/testthat/test-codelist.R
@@ -2,7 +2,7 @@
 cs <- countrycode::codelist
 pan <- countrycode::codelist_panel
 
-dest <- c('year', 'ar5', 'continent', 'currency', 'eu28', 'eurocontrol_pru',
+dest <- c('year', 'ar5', 'continent', 'currency', 'eu28', 'eu27', 'eurocontrol_pru',
           'eurocontrol_statfor', 'icao', 'icao.region', 'iso4217c',
           'iso4217n','region', 'region23', 'un.region.name', 'un.region.code',
           'un.regionsub.name', 'un.regionsub.code', 'telephone',
@@ -27,6 +27,15 @@ test_that('codelist has (roughly) correct dimensions', {
     expect_lt(ncol(cs), 650)
 })
 
+test_that('eu27 matches eu28 except for the United Kingdom', {
+    uk <- cs$country.name.en == 'United Kingdom'
+    expect_equal(sum(uk), 1)
+    expect_equal(cs$eu28[uk], 'EU')
+    expect_true(is.na(cs$eu27[uk]))
+    expect_equal(cs$eu27[!uk], cs$eu28[!uk])
+    expect_equal(sum(cs$eu27 == 'EU', na.rm = TRUE), 27)
+})
+
 # columns
 cols <- c('country.name.en.regex', 'country.name.en', 'iso3c', 'cowc', 'p4c', 'vdem')
 for (i in cols) {
@@ -46,7 +55,7 @@ test_that('codelist missing values', {
     }
 })
 
-cols <- setdiff(colnames(cs), c('dhs', 'eu28', 'un.regionintermediate.name',
+cols <- setdiff(colnames(cs), c('dhs', 'eu28', 'eu27', 'un.regionintermediate.name',
                                 'un.regionintermediate.code'))
 for (i in cols) {
     msg <- paste(i, 'has less than 50% missing observations')
@@ -89,7 +98,7 @@ test_that('codelist_panel has (roughly) correct dimensions', {
     expect_gt(nrow(pan), 25000)
     expect_lt(nrow(pan), 30000)
     expect_gt(ncol(pan), 40)
-    expect_lt(ncol(pan), 60)
+    expect_lt(ncol(pan), 61)
 })
 
 # columns
@@ -112,7 +121,7 @@ test_that('codelist_panel missing values', {
     }
 })
 
-cols <- setdiff(colnames(pan), c('dhs', 'eu28', 'un.regionintermediate.name',
+cols <- setdiff(colnames(pan), c('dhs', 'eu28', 'eu27', 'un.regionintermediate.name',
                                  'un.regionintermediate.code'))
 for (i in cols) {
     msg <- paste(i, 'has less than 50% missing observations')